split trigger_lists
