+++ AoE2ScenarioParser/datasets/trigger_lists/__init__.py
@@ -1,0 +1,37 @@
+from .action_type import *
+from .age import *
+from .attack_stance import *
+from .attribute import Attribute
+from .blast_level import *
+from .button_location import *
+from .charge_event import *
+from .charge_type import *
+from .color_mood import *
+from .combat_ability import *
+from .comparison import *
+from .damage_class import *
+from .difficulty_level import *
+from .diplomacy_state import *
+from .fog_visibility import *
+from .garrison_type import *
+from .hero_status_flag import *
+from .hotkey import *
+from .object_attribute import *
+from .object_class import *
+from .object_state import *
+from .object_type import *
+from .occlusion_mode import *
+from .operation import *
+from .panel_location import *
+from .projectile_smart_mode import *
+from .projectile_vanish_mode import *
+from .prrojectile_hit_mode import *
+from .secondary_game_mode import *
+from .technology_state import *
+from .terrain_restrictions import *
+from .time_unit import *
+from .unit_action import *
+from .unit_trait import *
+from .victorry_condition import *
+from .victory_timer_type import *
+from .visibility_state import *
